Pass data between pre, intra, post ThreadManagers
Currently, in each process spawned by bmap, the data handled by the pre,
intra, and post functions are passed between ThreadManager objects via
the ProcessHandler. It will be better to just have the results of each
step be piped directly to the next step. This commit adds this
functionality.

Note, that each sub-step (pre, intra, post) has a queue limit of
"preload", but this commit changes the limit to apply to the entire
pipeline.