Fix preload in pool_map
Currently, the preload sets the per-process number of items to handle at
a time and the total number of items sent to all pools at a time. The
total number should be larger to see any performance gains. This commit
adds a pool_preload (default 2 * n_procs * preload) argument to set the
total preload.